Code moved to utils folder for showLoader and hideLoader
